+++ lib/BicBucStriim/l10n.php
@@ -2,7 +2,7 @@
 /**
 * Localization helper
 */
-require_once 'langs.php';
+require_once __DIR__ . '/langs.php';
 
 class L10n extends ArrayObject
 {